fitting: Fit RGB2LMS matrix

+++ lib/spectra.cc
@@ -42,10 +42,10 @@ spectrum spectrum::operator+(const spectrum &other) {
     return res;
 }
 
-float spectrum::integrate() const {
+double spectrum::integrate() const {
     double res = std::accumulate(values.cbegin(), values.cend(), 0.0);
     res *= wl_step;
-    return static_cast<float>(res);
+    return res;
 }
